fix: upsert/update data ttl is integer

See merge request: !18

+++ src/main/java/com/volcengine/vikingdb/runtime/vector/model/request/UpdateDataRequest.java
@@ -23,7 +23,7 @@ public class UpdateDataRequest extends BaseCollectionRequest {
     private List<Map<String, Object>> data;
 
     @JsonProperty("ttl")
-    private Long ttl;
+    private Integer ttl;
 
     @JsonProperty("ignore_unknown_fields")
     private Boolean ignoreUnknownFields;
